WRP-2904

+++ app/shared/concept-edit/conceptEdit.js
@@ -669,7 +669,7 @@ angular.module('singleConceptAuthoringApp').directive('conceptEdit', function ($
           else {
             if (deletion) {
               selectInactivationReason('Concept', inactivateConceptReasons, inactivateAssociationReasons, scope.concept.conceptId, scope.concept, scope.branch, deletion).then(function (results) {
-                if (results.deletion) {
+                if (results.deletion && !results.reason) {
                   notificationService.sendMessage('Deleting Concept...');
                   snowowlService.deleteConcept(scope.concept.conceptId, scope.branch).then(function (response) {
                     if (response.status === 409) {
@@ -682,6 +682,10 @@ angular.module('singleConceptAuthoringApp').directive('conceptEdit', function ($
                       notificationService.sendMessage('Concept Deleted', 5000);
                     }
                   });
+                }
+                else{
+                   inactivationService.setParameters(scope.branch, scope.concept, results.reason.id, results.associationTarget, results.deletion);
+                  $rootScope.$broadcast('conceptEdit.inactivateConcept');
                 }
               });
             }
